refactor(server): delete App/McpServer/Skill services — Fixes EVE-316
