+++ crates/server/src/domains/skills/mod.rs
@@ -4,6 +4,7 @@
 
 use everruns_core::{Permission, Policy, Rule};
 
+pub mod archive;
 pub mod commands;
 pub mod queries;
 pub mod types;
